requirements.txt
sounddevice==0.5.5
vosk==0.3.45
PyYAML==6.0.2